requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
cachetools==5.5.0
    # via google-auth
certifi==2024.12.14
    # via requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via -r requirements.in
google-api-core==2.17.0
    # via google-api-python-client
google-api-python-client==2.116.0
    # via -r requirements.in
google-auth==2.25.2
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-auth-oauthlib==1.2.0
    # via -r requirements.in
googleapis-common-protos==1.62.0
    # via google-api-core
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.10
    # via requests
iniconfig==2.0.0
    # via pytest
mypy-extensions==1.0.0
    # via -r requirements.in
oauthlib==3.2.2
    # via requests-oauthlib
packaging==24.2
    # via pytest
pathspec==0.12.1
    # via -r requirements.in
platformdirs==4.3.6
    # via -r requirements.in
pluggy==1.5.0
    # via pytest
proto-plus==1.25.0
    # via -r requirements.in
protobuf==4.25.5
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pyparsing==3.2.0
    # via httplib2
pyproject-hooks==1.2.0
    # via -r requirements.in
pytest==7.4.4
    # via -r requirements.in
python-dateutil==2.8.2
    # via -r requirements.in
requests==2.32.3
    # via
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
schedule==1.2.1
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.3
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
